drag + code restructure
applyAttributes() now uses dicts instead of a wall of if statements. Drag function now set up. particles are no longer relative to the emitter and instead are just placed on screen space. Pos can now update properly and velocity can be passed to the particles.

Also found a chunk of code to do alpha, however I don't know if it's the right thing to implement with this project.