many small changes
did a bunch of small changes to account for the wiki being created. not sure if the notes in particles.py for each function are still 100% up to date, but it should be fine bcs the wiki is all set up now.

Only big change is that now instead of passing the veloType, it is automatically set up in the init of the particle emitter.