slight change to load func
Now it's gonna work regardless of if the notes are set as "" or not. Copying the data doesn't copy the spaces at the end of each transaction line so doing it this way makes it so that lines can be accessed regardless of if the notes are there or not.